Store modules by fixed key instead type object
